+++ forge/src/main/java/de/cech12/usefulhats/UsefulHatsForgeMod.java
@@ -1,17 +1,12 @@
 package de.cech12.usefulhats;
 
-import de.cech12.usefulhats.client.UsefulHatLayers;
 import de.cech12.usefulhats.compat.CuriosMod;
 import de.cech12.usefulhats.init.ModCreativeTabs;
 import de.cech12.usefulhats.init.ModItems;
 import de.cech12.usefulhats.platform.Services;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.InterModEnqueueEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
@@ -26,11 +21,6 @@ public class UsefulHatsForgeMod {
         ModCreativeTabs.TABS.register(modEventBus);
         //Configs
         CommonLoader.init();
-        //register render layers & models
-        DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> () -> {
-            FMLJavaModLoadingContext.get().getModEventBus().addListener(UsefulHatLayers::initLayers);
-            FMLJavaModLoadingContext.get().getModEventBus().addListener(UsefulHatLayers::initModels);
-        });
     }
 
     @SubscribeEvent