+++ forge/src/main/java/de/cech12/usefulhats/client/UsefulHatsForgeClientEvents.java
@@ -1,15 +1,20 @@
 package de.cech12.usefulhats.client;
 
 import de.cech12.usefulhats.Constants;
-import de.cech12.usefulhats.compat.CuriosMod;
+import de.cech12.usefulhats.client.compat.CurioClientCompat;
 import de.cech12.usefulhats.item.IGameOverlayRenderer;
 import de.cech12.usefulhats.platform.Services;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.color.item.ItemColors;
+import net.minecraft.client.model.geom.LayerDefinitions;
+import net.minecraft.client.model.geom.ModelLayerLocation;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.DyeableLeatherItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.EntityRenderersEvent;
 import net.minecraftforge.client.event.RegisterColorHandlersEvent;
 import net.minecraftforge.client.event.RegisterGuiOverlaysEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -20,12 +25,26 @@ import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 @Mod.EventBusSubscriber(modid = Constants.MOD_ID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class UsefulHatsForgeClientEvents {
 
+    public static final ModelLayerLocation USEFUL_HAT_LAYER = new ModelLayerLocation(new ResourceLocation(Constants.MOD_ID, "usefulhat_layer"), "main");
+
+    public static UsefulHatModel<LivingEntity> usefulHatModel = null;
+
     @SubscribeEvent
     public static void onClientRegister(FMLClientSetupEvent event) {
         //curio rendering
         if (Services.PLATFORM.isModLoaded(Constants.CURIOS_MOD_ID)) {
-            CuriosMod.setupClient();
+            CurioClientCompat.register();
         }
+    }
+
+    @SubscribeEvent
+    public static void initLayers(EntityRenderersEvent.RegisterLayerDefinitions event) {
+        event.registerLayerDefinition(USEFUL_HAT_LAYER, () -> UsefulHatModel.createLayer(LayerDefinitions.INNER_ARMOR_DEFORMATION, 0));
+    }
+
+    @SubscribeEvent
+    public static void initModels(EntityRenderersEvent.AddLayers event) {
+        usefulHatModel = new UsefulHatModel<>(Minecraft.getInstance().getEntityModels().bakeLayer(USEFUL_HAT_LAYER));
     }
 
     @SubscribeEvent