+++ forge/src/main/java/de/cech12/usefulhats/compat/CuriosMod.java
@@ -2,19 +2,14 @@ package de.cech12.usefulhats.compat;
 
 import de.cech12.usefulhats.Constants;
 import de.cech12.usefulhats.UsefulHatsEventUtils;
-import de.cech12.usefulhats.client.compat.CurioRenderer;
-import de.cech12.usefulhats.init.ModItems;
 import de.cech12.usefulhats.item.AbstractHatItem;
 import de.cech12.usefulhats.platform.Services;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.fml.InterModComms;
-import net.minecraftforge.registries.RegistryObject;
 import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.SlotTypeMessage;
 import top.theillusivec4.curios.api.SlotTypePreset;
-import top.theillusivec4.curios.api.client.CuriosRendererRegistry;
 import top.theillusivec4.curios.api.event.CurioChangeEvent;
 
 import java.util.List;
@@ -26,12 +21,6 @@ public class CuriosMod {
     public static void addHeadSlot() {
         if (Services.CONFIG.isCuriosAddHeadSlot()) {
             InterModComms.sendTo(Constants.CURIOS_MOD_ID, SlotTypeMessage.REGISTER_TYPE, () -> SlotTypePreset.HEAD.getMessageBuilder().build());
-        }
-    }
-
-    public static void setupClient() {
-        for (RegistryObject<Item> item : ModItems.ITEMS.getEntries()) {
-            CuriosRendererRegistry.register(item.get(), CurioRenderer::getInstance);
         }
     }